requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
msal>=1.28.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
azure-monitor-opentelemetry>=1.6.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
